Added docs for doing mass deletes with Model#destroy
Also, changed usage of destroy! to destroy for now.
I will add a section explaining the general difference
between bang and non-bang methods soon.

Since #create, #update and #save are all explained
in their non-bang version currently, it's better to
be consistent. Also, the non-bang version is more
safe to use and bangs should be used with care anyway.