Some corrections to the docs about nested conditions
We now only mention that property and relationship
*names* are available as keys in nested conditions.

While it's technically possible to use Property
objects as keys too, I'd think that the use for
that is maybe rather rare. We can always add it
later if we decide that it's worth going into that
much detail.

Using relationship objects as keys was confusing
and wrong to begin with. A future commit will add
some docs for constructing QueryPath objects,
which is the actual concept behind the idea.